+++ source/problems/posterior.cpp
@@ -8,7 +8,7 @@ Korali::Problem::Posterior::Posterior(void (*modelFunction) (double*, double*), 
 double Korali::Problem::Posterior::evaluateFitness(double* sample)
 {
   double posterior = Korali::Problem::Likelihood::evaluateFitness(sample);
-  for (int i = 0; i < _parameterCount; i++) posterior += log(_parameters[i]->getDensity(sample[i]));
+  for (size_t i = 0; i < _parameterCount; i++) posterior += log(_parameters[i]->getDensity(sample[i]));
 
   return posterior;
 }
